backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==2.5.1
Flask-Migrate==4.0.4
python-dotenv==1.0.1
PyMySQL==1.1.0
PyJWT==2.8.0
bcrypt==4.0.1
requests==2.32.5
celery==5.3.1
redis==4.5.5
Flask-Mail==0.9.1
Flask-Cors==3.0.10
Flask-Limiter==2.6.0
pytest==7.4.0
Flask-Admin==1.6.0
SQLAlchemy==1.4.49
